处理美客多订单拉取问题【临时上线】21

+++ erp-sdk/erp-sdk-oms-mercado/src/main/java/com/sdk/oms/mercado/dto/mercado/cost/SendersBean.java
@@ -33,7 +33,7 @@ public class SendersBean {
         return userId;
     }
 
-    public void setUserId(int userId) {
+    public void setUserId(long userId) {
         this.userId = userId;
     }
