+++ erp-sdk/erp-sdk-oms-mercado/src/main/java/com/sdk/oms/mercado/dto/mercado/shipment/ShippingAddressBean.java
@@ -72,7 +72,7 @@ public class ShippingAddressBean {
         return addressId;
     }
 
-    public void setAddressId(int addressId) {
+    public void setAddressId(long addressId) {
         this.addressId = addressId;
     }
